+++ src/objects/ad-account-targeting-unified.js
@@ -76,7 +76,6 @@ export default class AdAccountTargetingUnified extends AbstractCrudObject {
       credit: 'CREDIT',
       employment: 'EMPLOYMENT',
       housing: 'HOUSING',
-      issues_elections_politics: 'ISSUES_ELECTIONS_POLITICS',
       none: 'NONE',
     });
   }