+++ src/objects/ad-creative.js
@@ -50,6 +50,7 @@ export default class AdCreative extends AbstractCrudObject {
       instagram_story_id: 'instagram_story_id',
       interactive_components_spec: 'interactive_components_spec',
       link_deep_link_url: 'link_deep_link_url',
+      link_destination_display_url: 'link_destination_display_url',
       link_og_id: 'link_og_id',
       link_url: 'link_url',
       messenger_sponsored_message: 'messenger_sponsored_message',