+++ src/objects/ad-set.js
@@ -225,7 +225,6 @@ export default class AdSet extends AbstractCrudObject {
       credit: 'CREDIT',
       employment: 'EMPLOYMENT',
       housing: 'HOUSING',
-      issues_elections_politics: 'ISSUES_ELECTIONS_POLITICS',
       none: 'NONE',
     });
   }