+++ src/objects/application.js
@@ -114,6 +114,7 @@ export default class Application extends AbstractCrudObject {
       smart_login_menu_icon_url: 'smart_login_menu_icon_url',
       social_discovery: 'social_discovery',
       subcategory: 'subcategory',
+      suggested_events_setting: 'suggested_events_setting',
       supported_platforms: 'supported_platforms',
       supports_apprequests_fast_app_switch: 'supports_apprequests_fast_app_switch',
       supports_attribution: 'supports_attribution',