+++ src/objects/business-unit.js
@@ -81,6 +81,16 @@ export default class BusinessUnit extends AbstractCrudObject {
     );
   }
 
+  getConversionPaths (fields: Array<string>, params: Object = {}, fetchFirstPage: boolean = true): Cursor | Promise<*> {
+    return this.getEdge(
+      AbstractObject,
+      fields,
+      params,
+      fetchFirstPage,
+      '/conversion_paths'
+    );
+  }
+
   getCustomBreakdowns (fields: Array<string>, params: Object = {}, fetchFirstPage: boolean = true): Cursor | Promise<*> {
     return this.getEdge(
       AbstractObject,