+++ src/objects/work-user-frontline.js
@@ -16,6 +16,7 @@ import {AbstractCrudObject} from './../abstract-crud-object';
 export default class WorkUserFrontline extends AbstractCrudObject {
   static get Fields () {
     return Object.freeze({
+      has_access: 'has_access',
       is_frontline: 'is_frontline',
     });
   }